requirements.txt
eventlet
bcrypt
joblib==1.5.1
numpy==2.3.1
pandas==2.3.1
Flask==3.1.1
scikit-learn==1.7.1
datasets==4.0.0
gunicorn
markdown2==2.4.4
pytz==2025.2
Flask-SocketIO==5.5.1
scipy==1.16.0
plotly==6.2.0
matplotlib==3.10.5
PuLP==3.2.2
psycopg2-binary
python-dotenv
Jinja2
flask-login
requests
pymysql
transformers
bleach
flask-cors
flask-sqlalchemy
scikit-learn
flask-limiter
typing
xxhash
Flax
redis>=3.0.0
uuid
google.generativeai
genai
pymongo
mysql.connector
asyncio

